Main v2 prod

+++ frontend/src/composables/api/user.service.ts
@@ -1,5 +1,5 @@
 import {Api} from "@/composables/api/routesApi.ts";
-import { createUser } from '@/dto/user.dto';
+import { User } from '@/dto/user.dto';
 
 const baseUrl = import.meta.env.VITE_APP_API_URL;
 
@@ -13,7 +13,7 @@ export const UserService = () => {
         if( token === null) {
             return handler(401);
         }
-        return await fetch(baseUrl + Api.user + `${id}` + params, {
+        return await fetch(baseUrl + Api.user + `/${id}` + params, {
             method: "GET",
             credentials: 'include',
             headers: {
@@ -27,34 +27,17 @@ export const UserService = () => {
     const updateUser = async (id: string, data: any, handler: Function) => {
         const token = localStorage.getItem('auth_token');
         if (token === null) {
-            return handler(401);
+            handler(new Response(null, { status: 401 }));
         }
-    
-        try {
-            const response = await fetch(baseUrl + Api.user + `${id}`, {
-                method: "PATCH",
-                credentials: 'include',
-                headers: {
-                    "Content-Type": "application/json",
-                    Authorization: `Bearer ${token}`,
-                },
-                body: JSON.stringify(data),
-            });
-            
-            if (!response.ok) {
-                const errorText = await response.text();
-                try {
-                    const errorJson = JSON.parse(errorText);
-                    throw errorJson;
-                } catch (err) {
-                    throw errorText;
-                }
-            }
-    
-            return handler(response);
-        } catch (error) {
-            throw error;
-        }
+        await fetch(baseUrl + Api.user + `/${id}`, {
+            method: "PATCH",
+            credentials: 'include',
+            headers: {
+                "Content-Type": "application/json",
+                Authorization: `Bearer ${token}`,
+            },
+            body: JSON.stringify(data),
+        }).then(res => handler(res));
     }    
     const getUsers = async (handler:Function) => {
         const token = localStorage.getItem('auth_token');
@@ -69,10 +52,27 @@ export const UserService = () => {
             }
         }).then(res => handler(res));
     }
-
+    const checkUser = async(UserId: string, password: string) => {
+        const token = localStorage.getItem('auth_token');
+        if(token === null) throw new Error('Error while identifying user');
+        return await fetch(baseUrl + `${Api.user}` + `/check`, {
+            method: 'POST',
+            headers: {
+                "Content-Type": "application/json",
+                Authorization: `Bearer ${token}`,
+            },
+            body: JSON.stringify({ UserId, password })
+        });
+    }
     const deleteUser = async (id: string) => {
-        const response = await fetch(`${baseUrl}${Api.user}${id}`, {
+        const token = localStorage.getItem('auth_token');
+        if(token === null) throw new Error('Error while deleting user');
+        const response = await fetch(baseUrl + `${Api.user}` + `/${id}`, {
             method: 'DELETE',
+            headers: {
+                "Content-Type": "application/json",
+                Authorization: `Bearer ${token}`,
+            },
         });
         if (!response.ok) {
             const errorData = await response.json();
@@ -82,6 +82,8 @@ export const UserService = () => {
     }
 
     const deleteBatchUsers = async (users: User[]) => {
+        const token = localStorage.getItem('auth_token');
+        if(token === null) throw new Error('Error while deleting users');
         const nonAdminUsers = users.filter(user => user.role !== 'admin');
         const nonAdminIds = nonAdminUsers.map(user => user.id);
         
@@ -92,7 +94,8 @@ export const UserService = () => {
         const response = await fetch(`${baseUrl}${Api.user}`, {
             method: 'DELETE',
             headers: {
-                'Content-Type': 'application/json'
+                'Content-Type': 'application/json',
+                Authorization: `Bearer ${token}`,
             },
             body: JSON.stringify({ usersId: nonAdminIds.join(',') })
         });
@@ -105,11 +108,11 @@ export const UserService = () => {
         return response;
     }
 
-    const createUser = async (bodyRequest: createUser): Promise<{sessionId: string}> => {
+    const createUser = async (bodyRequest: any): Promise<Response> => {
         try {
             const token = localStorage.getItem('auth_token');
             if(token === null) throw new Error('Error while getting orders');
-            const response = await fetch(baseUrl + Api.user, {
+            return await fetch(baseUrl + Api.user, {
                 method: 'POST',
                 headers: {
                     'Content-Type': 'application/json',
@@ -117,21 +120,9 @@ export const UserService = () => {
                 },
                 body: JSON.stringify(bodyRequest)
             });
-
-            if (!response.ok) {
-                const errorText = await response.text();
-                try {
-                    const errorJson = JSON.parse(errorText);
-                    throw errorJson;
-                } catch (err) {
-                    throw errorText;
-                }
-            }
-        
-            return await response.json();
         } catch (err) {
             throw err;
         }
     }
-    return { getUserById, getUsers, updateUser, deleteUser, deleteBatchUsers, createUser }
+    return { checkUser, getUserById, getUsers, updateUser, deleteUser, deleteBatchUsers, createUser }
 }
